feat(object): Update object base on object sequence diagram

+++ application/sources/app/tiny_rex_game/game_object.h
@@ -9,7 +9,7 @@ typedef struct {
     bool visible;
     uint8_t state;
     uint8_t speed;
-    int8_t x, y;
+    int16_t x, y;
     uint8_t action_image;
     uint8_t level;
 } game_object_t;
